Catch errors

+++ get_following.py
@@ -31,25 +31,32 @@ print(df)
 
 # Create dataframe to store retrieved data in
 data = pd.DataFrame()
+data["user_id"] = ""
+data["following_list"] = ""
+data["followers_list"] = ""
+
 
 # Use Twitter API to get IDs of all users followed by and following the users in the dataset
 for idx, user_id in enumerate(df["user_id"]):
    
     # Print progress
-    #if (idx % 10 == 0):
-        #print(str(round(idx*100/len(df["user_id"]))) + "% progress", end="\r", flush=True)
     print(idx, "of", df_length, "rows processed", end="\r", flush=True)
+
+    # Keep track whether current user is protected
+    protected = False
 
     # We attempt to connect up to 50 times in case of disconnects, resets etc.
     for attempt in range(50):
         # Get following and followers for user
         try:
+            #print("try", idx, "attempt", attempt)
             following = api.friends_ids(user_id=user_id)
             followers = api.followers_ids(user_id=user_id)
         # Handle exception in case the user does not exist or user is protected
         except tweepy.error.TweepError as e:
             if (e.reason == "Not authorized."):
-                continue
+                protected = True
+                break
             print("Error: ", e)
         # Handle exception in case of connection timeouts or resets
         except (Timeout, SSLError, ReadTimeoutError, ConnectionError) as e:
@@ -60,16 +67,20 @@ for idx, user_id in enumerate(df["user_id"]):
         else:
             break
 
+    # If user is protected, skip this user
+    if (protected):
+        continue
+
     # Store retrieved data in dataframe, use list of lists because column needs to have the same length
-    data["user_id"] = user_id
-    data["following_list"] = [following]
-    data["followers_list"] = [followers]
+    data.at[idx,"user_id"] = user_id
+    data.at[idx,"following_list"] = [following]
+    data.at[idx,"followers_list"] = [followers]
 
     if (idx % 10 == 0):
         data.to_csv("data_following_progress.csv", index=None, header=True)
 
     # Wait one minute to avoid rate limit
-    time.sleep(63)
+    time.sleep(61)
 
 # Merge retrieved data with the dataset
 data.to_csv("data_following_scraped.csv", index=None, header=True)
